// src/models/f1/car-frame.ts
export const CAR_FRAME = {
// ...
  parts: {
    "front-wing": {
      FrontWingLivery: {
        min: [-3.012, 0.129, -0.595],
        max: [-2.322, 0.309, 0.595]
      },
      material_0: {
        min: [-2.578, 0.283, -0.878],
        max: [-2.339, 0.33, 0.878]
      },
      FrontEndplateLivery: {
        min: [-2.759, 0.104, -0.882],
        max: [-2.094, 0.394, 0.882]
      },
      Carbon: {
        min: [-2.88, 0.142, -0.004],
        max: [-2.696, 0.183, 0.004]
      }
    },
    "rear-wing": {
      RearWingLivery: {
        min: [1.943, 0.815, -0.566],
        max: [2.327, 0.935, 0.566]
      },
      RearEndplateLivery: {
        min: [1.932, 0.369, -0.577],
        max: [2.456, 0.936, 0.577]
      },
      Carbon: {
        min: [1.793, 0.434, -0.07],
        max: [2.249, 0.964, 0.07]
      }
    },
// ...
    "front-left-wheel": {
      Carbon: {
        min: [-1.921, 0.149, 0.526],
        max: [-1.479, 0.59, 0.614]
      },
      Cover: {
        min: [-1.954, 0.115, 0.598],
        max: [-1.446, 0.624, 0.947]
      },
      Tyre: {
        min: [-2.069, 0, 0.59],
        max: [-1.331, 0.739, 0.954]
      },
      Metal: {
        min: [-1.718, 0.352, 0.921],
        max: [-1.682, 0.387, 0.953]
      }
    },
// ...
    "rear-left-wheel": {
      Carbon: {
        min: [1.479, 0.149, 0.517],
        max: [1.921, 0.59, 0.536]
      },
      Cover: {
        min: [1.446, 0.115, 0.522],
        max: [1.954, 0.624, 0.948]
      },
      Tyre: {
        min: [1.331, 0, 0.513],
        max: [2.069, 0.739, 0.955]
      },
      Metal: {
        min: [1.682, 0.352, 0.921],
        max: [1.718, 0.387, 0.954]
      }
    },
// ...
  }
} as const;